Use macro instead of runtime AST manipulation.
Remove Flexer.CodeBuilder
